+++ src/main/java/roomescape/dto/LoginRequest.java
@@ -1,15 +1,13 @@
-package roomescape.member;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record MemberRequest(
-        @NotBlank(message = "이름은 필수입니다.")
-        String name,
-
+public record LoginRequest(
         @NotBlank(message = "이메일은 필수입니다.")
         String email,
 
         @NotBlank(message = "비밀번호는 필수입니다.")
         String password
-){
+) {
+
 }
