+++ src/main/java/roomescape/dto/MemberResponse.java
@@ -1,3 +1,3 @@
-package roomescape.member;
+package roomescape.dto;
 public record MemberResponse(Long id, String name, String email) {
 }
