+++ src/main/java/roomescape/member/MemberController.java
@@ -1,7 +1,5 @@
 package roomescape.member;
 
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.ResponseEntity;
@@ -9,7 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.util.JwtUtil;
+import roomescape.dto.LoginRequest;
+import roomescape.dto.MemberRequest;
+import roomescape.dto.MemberResponse;
 
 import java.net.URI;
